pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.0"
description = ""
authors = ["Developer <dev@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
Django = "==5.2"
psycopg2-binary = "2.9.10"
drf-yasg = "1.21.11"
boto3 = "1.40.45"
django-storages = "1.14.6"
gunicorn = "23.0.0"
django-environ = "0.12.0"
django-filter = "25.2"
shortuuid = "1.0.13"
types-python-dateutil = "2.9.0.20250822"
django-cors-headers = "4.9.0"
black = "25.9.0"
django-extensions = "4.1"
python-jose = "3.5.0"
pydantic = "1.10.24"
djangorestframework = "3.16.1"
celery = { extras = ["redis"], version = "5.3.1" }
django-celery-beat = "2.8.1"
redis = "6.4.0"
flower = "2.0.1"
types-redis = "4.6.0.20241004"
pandas = "2.3.3"
openpyxl = "3.1.5"
spacy = "3.8.7"
openai = "2.3.0"
django-model-utils = "5.0.0"
pyphen = "0.17.2"
stripe = "5.5.0"
sendgrid = "6.12.5"
# External algorithm library (reference removed for anonymization)
# algorithm-library = {git = "git@github.com:example/algorithm.git", tag = "v1.0.0" }
coreapi = "2.3.3"
hubspot-api-client = "12.0.0"
cachetools = "6.2.0"
# See: https://stackoverflow.com/questions/78634235/numpy-dtype-size-changed-may-indicate-binary-incompatibility-expected-96-from
numpy = "2.3.3"


[tool.poetry.group.dev.dependencies]
mypy = "1.18.2"
django-stubs = "5.2.6"
flake8 = "7.3.0"
isort = "6.1.0"
drf-yasg-stubs = "0.1.4"
djangorestframework-stubs = "3.14.2"
pre-commit = "4.3.0"
types-redis = "4.6.0.20241004"
pytest = "8.4.2"
pytest-cov = "6.3.0"
coverage = "7.10.7"
pytest-django = "4.11.1"
pytest-mock = "3.15.1"
types-cachetools = "5.5.0.20240820"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
    (
        \.eggs|
        \.git|
        \.hg|
        \.mypy_cache|
        \.nox|
        \.tox|
        \.venv|
        \.svn|
        _build|
        buck-out|
        build|
        dist|
        settings\.py|
        migrations|
        snapshots
    )/
'''

[tool.tox]
legacy_tox_ini = """
[flake8]
ignore =
    # To be PEP-8 compliant, Black splits lines before binary operators
    W503,
    # For the same reason, Black enforces whitespace around slice operators
    E203,
    # Do not enforce newlines before class docstrings
    D203
exclude =
    .git
    __pycache__,
    dist,
    build,
    provision,
    venv,
    settings.py,
    migrations
# Black defaults to 88 and has some pretty good reasons for it
# a.k.a: Raymond Hettinger says so:
# https://www.youtube.com/watch?v=wf-BqAjZb8M&feature=youtu.be&t=260
max-line-length=88
"""

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
skip_glob = '*migrations*,*snapshot*'
